chore(scripts): Split revalidate-stale-quarantines.test.ts (SMI-5865)
The file was 507 lines, over the pre-commit check-file-length.mjs
500-line hard-fail limit (grandfathered pending this split during SMI-5849).
Split into two concern-scoped siblings along the file's existing describe
groupings: revalidate-stale-quarantines.outcomes.test.ts (outcome-branch
tests -- parse-failed, repo-gone, kept-security, sibling-recovered,
fetch-error, 345 lines, 11 tests) and revalidate-stale-quarantines.guards.test.ts
(CAS/regression guard tests -- cas-skipped, error, loadCandidates pagination,
E4/E9 regression pins, 307 lines, 7 tests). All 18 original it() blocks
preserved exactly, zero coverage lost.

Updated revalidate-stale-quarantines.requarantine.test.ts's docstring, which
referenced the now-deleted filename. Cleared the grandfather entry in
scripts/check-file-length.ignore per that file's documented historical-note
format.

Refs SMI-5865
